ai-microservice/ai_classifier.py
import requests
from io import BytesIO
from PIL import Image
# NOTE: In a real implementation, you would import heavy libraries here:
# from ultralytics import YOLO 
# import whisper
# import librosa

# --- Placeholder for Loaded Models ---
# Global variables to hold the loaded models (ensures models are loaded only once)
IMAGE_MODEL = None
NLP_MODEL = None

def load_models():
    """
    Loads and initializes the necessary AI models into memory.
    This function should be called once when the microservice starts.
    """
    global IMAGE_MODEL, NLP_MODEL
    print("AI_CLASSIFIER: Starting model loading process...")
    
    # 1. Image Model (YOLOv8/Detectron2) - Load weights and configuration
    # IMAGE_MODEL = YOLO('pothole_detection_v8.pt') 
    IMAGE_MODEL = "Mock_YOLOv8_Loaded" # Placeholder
    print("AI_CLASSIFIER: Image detection model loaded.")

    # 2. NLP/STT Model (Whisper/Gemini-2.5-Flash-Preview-09-2025) - Load STT and NLP pipeline
    # NLP_MODEL = whisper.load_model("base") 
    NLP_MODEL = "Mock_Whisper_NLP_Loaded" # Placeholder
    print("AI_CLASSIFIER: Speech/NLP model loaded.")

    print("AI_CLASSIFIER: All models initialized successfully.")

# --- Helper Functions for Inference ---

def _analyze_image_for_severity(image_url):
    """
    Simulates downloading the image and running object detection (YOLOv8).
    In real life, this would:
    1. Download image_url (requests.get).
    2. Run IMAGE_MODEL.predict() to get object bounding boxes (pothole, garbage).
    3. Calculate severity based on object size relative to image frame.
    """
    # Placeholder Logic: Always detect a 'pothole' and assign a medium score unless specified
    if "garbage" in image_url.lower():
        issue_type = "garbage"
        severity_score = 4
    elif "pothole" in image_url.lower():
        issue_type = "pothole"
        severity_score = 5  # High risk default
    else:
        issue_type = "street_light"
        severity_score = 2

    print(f"AI_CLASSIFIER: Analyzed image. Type: {issue_type}, Score: {severity_score}")

    return {
        "issueType": issue_type,
        "severityScore": severity_score,
        "tags": [issue_type.replace('_', '-'), "ai-classified"]
    }

def _process_audio_for_structure(audio_url, user_description):
    """
    Simulates downloading the audio, running STT (Whisper), and structuring (NLP).
    In real life, this would:
    1. Download audio_url (requests.get).
    2. Run NLP_MODEL (Whisper) to transcribe the audio.
    3. Run an NLP pipeline on the transcription to extract entities (issue, location).
    """
    # Mock STT/NLP: We'll pretend the audio contained detailed info and transcription succeeded.
    transcribed_text = f"AUTO_TRANSCRIPT: The voice note was about a deep water leak near the {user_description} area."
    
    # Mock NLP extraction logic
    if "leak" in transcribed_text.lower() or "water" in transcribed_text.lower():
        issue_type = "water_leak"
        severity_score = 4
        tags = ["water-leak", "high-priority"]
    else:
        issue_type = "other"
        severity_score = 3
        tags = ["voice-report"]

    print(f"AI_CLASSIFIER: Processed audio. Type: {issue_type}")

    return {
        "issueType": issue_type,
        "severityScore": severity_score,
        "tags": tags,
        "description": transcribed_text # Return the clean transcription to the main backend
    }

# --- Main Public Functions ---

def classify_image(image_url, user_description):
    """
    Main function for image reports.
    """
    if IMAGE_MODEL is None:
        return {"error": "Image model not loaded"}, 500
    
    # 1. Run image analysis
    results = _analyze_image_for_severity(image_url)

    # 2. Combine results and return
    return results

def classify_audio(audio_url, user_description):
    """
    Main function for audio reports.
    """
    if NLP_MODEL is None:
        return {"error": "NLP model not loaded"}, 500
    
    # 1. Run audio processing
    results = _process_audio_for_structure(audio_url, user_description)

    # 2. Combine results and return
    return results

# Initialize models upon module load (should be called by the Flask app start)
# load_models()